Various fixes

+++ manjaro_i3/DriveMounting.py
@@ -2,34 +2,46 @@ import os
 import FileManaging
 import subprocess
 
-subprocess.run([""], check=True)
-
 def writeToFstabAndMount(mountString, mountPoint):
     with open("/etc/fstab", "a") as fstab:
         fstab.write(mountString + "\n")
-    os.rmdir(f'/mnt/{mountPoint}')
+
+    if os.path.isdir("/mnt/" + mountPoint):  
+        os.rmdir(f'/mnt/{mountPoint}')
+    
     os.mkdir(f'/mnt/{mountPoint}')
 
 def getExt4MountString(drive, mountPoint):
-    return f'UUID={drive}    /mnt/{mountPoint}    ext4    defaults    0    1'
+    remoteDrive = f'UUID={drive}    '
+    localMount = f'/mnt/{mountPoint}    '
+    filetype = "ext4    "
+    idStrings = "defaults    0    1"
+
+    return remoteDrive + localMount + filetype + idStrings
 
 def getCifsMountString(drive, mountPoint, nasIpAddress):
-    return f'//{nasIpAddress["serverIp"]}/{drive} /mnt/{mountPoint} cifs credentials={nasIpAddress["credentials"]},uid=1000,gid=1000 0 0'
+    remoteDrive = f'//{nasIpAddress}/{drive} '
+    localMount = f'/mnt/{mountPoint} '
+    filetype = "cifs "
+    credentials = f'credentials={nasIpAddress}'
+    idStrings = ",uid=1000,gid=1000 0 0"
+
+    return remoteDrive + localMount + filetype + credentials + idStrings
 
 def getIscsiMountString(drive, mountPoint, nasIpAddress):
-    subprocess.run([
-        "sudo", 
-        "iscsiadm", 
-        "--mode", 
-        "node", 
-        "--targetname", 
-        nasIpAddress["targetName"],
-        "--portal",
-        nasIpAddress["serverIp"],
-        "--login",
-        ],
-        check=True)    
-    subprocess.run(["sudo", "mkfs.ext4", "/dev/"+{mountPoint}], check=True)
+    # subprocess.run([
+    #     "sudo", 
+    #     "iscsiadm", 
+    #     "--mode", 
+    #     "node", 
+    #     "--targetname", 
+    #     nasIpAddress["targetName"],
+    #     "--portal",
+    #     nasIpAddress["serverIp"],
+    #     "--login",
+    #     ],
+    #     check=True)    
+    # subprocess.run(["sudo", "mkfs.ext4", "/dev/" + {mountPoint}], check=True)
     return f'/dev/{drive} /mnt/{mountPoint} ext4 _netdev,rw 0 0'
 
 def setupSmbCredentials(smbcredentials):
@@ -46,16 +58,16 @@ def mountDrives(drivesYaml, serverConfig):
     setupSmbCredentials(serverConfig["smbcredentials"])
 
     for drive in drives["ext4"]:
-        mountString = getExt4MountString(drive["drive"], drive["mountpoint"])
-        writeToFstabAndMount(mountString, drive["mountpoint"])
+        mountString = getExt4MountString(drive["drive"], drive["mountPoint"])
+        writeToFstabAndMount(mountString, drive["mountPoint"])
 
     for drive in drives["cifs"]:
-        mountString = getCifsMountString(drive["drive"], drive["mountpoint"], serverConfig["nasIpAddress"])
-        writeToFstabAndMount(mountString, drive["mountpoint"])
+        mountString = getCifsMountString(drive["drive"], drive["mountPoint"], serverConfig["nasIpAddress"])
+        writeToFstabAndMount(mountString, drive["mountPoint"])
 
-    for drive in drives["iscsi"]:
-        mountString = getIscsiMountString(drive["drive"], drive["mountpoint"], serverConfig["nasIpAddress"])
-        writeToFstabAndMount(mountString, drive["mountpoint"])
+    # for drive in drives["iscsi"]:
+    #     mountString = getIscsiMountString(drive["drive"], drive["mountPoint"], serverConfig["nasIpAddress"])
+    #     writeToFstabAndMount(mountString, drive["mountPoint"])
         
-        subprocess.run(["systemctl", "daemon-reload"], check=True)
-        subprocess.run(["sudo", "mount", "-a"], check=True)
+    #     subprocess.run(["systemctl", "daemon-reload"], check=True)
+    #     subprocess.run(["sudo", "mount", "-a"], check=True)
